Refactor protocol wizard.
