+++ apps/admin-portal/src/components/applications/wizard/index.ts
@@ -21,4 +21,3 @@ export * from "./general-settings-wizard-form";
 export * from "./oauth-protocol-settings-wizard-form";
 export * from "./protocol-selection-wizard-form";
 export * from "./wizard-summary";
-export * from "./application-add-protocol-wizard";
