+++ apps/admin-portal/src/pages/application-template.tsx
@@ -21,7 +21,7 @@ import React, { FunctionComponent, ReactElement, SyntheticEvent, useEffect, useS
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { Divider } from "semantic-ui-react";
-import { ApplicationCreateWizard, ApplicationProtocolAddWizard } from "../components";
+import { ApplicationCreateWizard } from "../components";
 import { CustomApplicationTemplate } from "../components/applications/meta";
 import { ApplicationTemplateIllustrations, EmptyPlaceholderIllustrations } from "../configs";
 import { history } from "../helpers";
@@ -42,8 +42,8 @@ export const ApplicationTemplateSelectPage: FunctionComponent<{}> = (): ReactEle
     const applicationTemplates: ApplicationTemplateListItemInterface[] = useSelector(
         (state: AppState) => state.application.templates);
 
-    const [ showWizard, setShowWizard ] = useState<boolean>(false);
-    const [ selectedTemplate, setSelectedTemplate]  = useState<ApplicationTemplateListItemInterface>(null);
+    const [showWizard, setShowWizard] = useState<boolean>(false);
+    const [selectedTemplate, setSelectedTemplate] = useState<ApplicationTemplateListItemInterface>(null);
     const [
         isApplicationTemplateRequestLoading,
         setApplicationTemplateRequestLoadingStatus
@@ -63,7 +63,7 @@ export const ApplicationTemplateSelectPage: FunctionComponent<{}> = (): ReactEle
             .finally(() => {
                 setApplicationTemplateRequestLoadingStatus(false);
             });
-    }, [ applicationTemplates ]);
+    }, [applicationTemplates]);
 
     /**
      * Handles back button click.
@@ -130,7 +130,7 @@ export const ApplicationTemplateSelectPage: FunctionComponent<{}> = (): ReactEle
                                         image={ EmptyPlaceholderIllustrations.newList }
                                         imageSize="tiny"
                                         title={ t("devPortal:components.templates.emptyPlaceholder.title") }
-                                        subtitle={ [ t("devPortal:components.templates.emptyPlaceholder.subtitles") ] }
+                                        subtitle={ [t("devPortal:components.templates.emptyPlaceholder.subtitles")] }
                                     />
                                 ) }
                                 tagsSectionTitle={ t("common:technologies") }
@@ -139,11 +139,11 @@ export const ApplicationTemplateSelectPage: FunctionComponent<{}> = (): ReactEle
                     )
                     : <ContentLoader dimmer/>
             }
-            <Divider hidden />
+            <Divider hidden/>
             <div className="custom-templates">
                 <TemplateGrid<ApplicationTemplateListItemInterface>
                     type="application"
-                    templates={ [ CustomApplicationTemplate ] }
+                    templates={ [CustomApplicationTemplate] }
                     templateIcons={ ApplicationTemplateIllustrations }
                     heading={ t("devPortal:components.applications.templates.manualSetup.heading") }
                     subHeading={ t("devPortal:components.applications.templates.manualSetup.subHeading") }
@@ -159,7 +159,7 @@ export const ApplicationTemplateSelectPage: FunctionComponent<{}> = (): ReactEle
                             image={ EmptyPlaceholderIllustrations.newList }
                             imageSize="tiny"
                             title={ t("devPortal:components.templates.emptyPlaceholder.title") }
-                            subtitle={ [ t("devPortal:components.templates.emptyPlaceholder.subtitles") ] }
+                            subtitle={ [t("devPortal:components.templates.emptyPlaceholder.subtitles")] }
                         />
                     ) }
                     tagsSectionTitle={ t("common:technologies") }
@@ -167,19 +167,13 @@ export const ApplicationTemplateSelectPage: FunctionComponent<{}> = (): ReactEle
             </div>
             {
                 showWizard && (
-                    <ApplicationProtocolAddWizard
+                    <ApplicationCreateWizard
                         title={ selectedTemplate?.name }
                         subTitle={ selectedTemplate?.description }
                         closeWizard={ (): void => setShowWizard(false) }
                         template={ selectedTemplate }
                         addProtocol={ false }
                     />
-                    // <ApplicationCreateWizard
-                    //     title={ selectedTemplate?.name }
-                    //     subTitle={ selectedTemplate?.description }
-                    //     closeWizard={ (): void => setShowWizard(false) }
-                    //     template={ selectedTemplate }
-                    // />
                 )
             }
         </PageLayout>
